Fix memset bug in powerspectrum: it makes no difference because both double and int64_t are 8 bytes!

+++ libgadget/powerspectrum.c
@@ -34,7 +34,7 @@ void powerspectrum_zero(struct _powerspectrum * PowerSpectrum)
 {
     memset(PowerSpectrum->kk, 0, sizeof(double) * PowerSpectrum->nalloc);
     memset(PowerSpectrum->Power, 0, sizeof(double) * PowerSpectrum->nalloc);
-    memset(PowerSpectrum->Nmodes, 0, sizeof(double) * PowerSpectrum->nalloc);
+    memset(PowerSpectrum->Nmodes, 0, sizeof(int64_t) * PowerSpectrum->nalloc);
     PowerSpectrum->Norm = 0;
 }
